vertexbuflayout warning fix

+++ src/System/Utils/VertexBufferLayout.cpp
@@ -4,7 +4,7 @@ namespace Utils
 {
 	void VertexBufferLayout::AddElementLayoutFloat( GLuint count, bool normalized )
 	{
-		layoutElments.push_back( { (GLint)count, GL_FLOAT, normalized, count * sizeof( GLfloat ) } );
+		layoutElments.push_back( { (GLint)count, GL_FLOAT, normalized, (GLuint)count * sizeof( GLfloat ) } );
 		stride += layoutElments.back().size;
 	}
 }
